+++ .eslintrc.js
@@ -11,6 +11,7 @@ module.exports = {
     '@typescript-eslint/ban-ts-comment': 'off',
     '@typescript-eslint/no-explicit-any': 'off',
     '@typescript-eslint/no-non-null-assertion': 'off',
+    '@typescript-eslint/no-unused-vars': 'off',
   },
   // https://eslint.org/docs/rules/no-undef#nodejs
   env: {